refactor(address): clean up shared address autocomplete
Follow-up cleanup for the FU-003 shared address autocomplete.

- Share the `ResolvedAddress` type from `geocoder.d.ts` instead of duplicating it.
- Remove unreachable `try/catch` blocks from `useAddressSuggestions`, since the geocoder helpers already handle errors internally.
- Use `magicKey` as the autocomplete option key to avoid duplicate React keys.
- Make `validationMessages` optional so call sites only provide the messages they use.
- Clean up leftover diff markers and update an outdated comment.
